backend/requirements.dev.txt
# Required
Django~=4.1.7
djangorestframework~=3.14.0
django-filter~=22.1
django-extensions~=3.2.1
drf-extensions~=0.7.1
django-cors-headers~=3.14.0
psycopg2~=2.9.5
django-cachalot~=2.5.3
django-redis~=5.2.0
django-bulk-sync~=3.3.0
django-storages~=1.13.2
boto3~=1.26.98
joblib~=1.2.0
Pillow~=9.4.0
djangorestframework-simplejwt~=5.2.2
dj-rest-auth[with_social]~=2.2.7
django-phonenumber-field~=7.0.2
phonenumbers~=8.13.7

# Dev
black~=23.1.0
flake8~=6.0.0
flake8-isort~=6.0.0
isort~=5.12.0
django-debug-toolbar~=3.8.1
drf-yasg~=1.21.5
tox~=4.4.7
mypy~=1.1.1
coverage~=7.2.2
pylint-django~=2.5.3
pre-commit~=3.2.0
django-coverage-plugin~=3.0.0
